Module evtc::buff

Buff bindings & utilities.

Buffs are temporary “effects” applied to agents. Some buffs modify the attributes of the destination agent. They can be positive like Boons, negative like Conditions or mixed. Some buffs do not do anything themselves and are simply used to as markers to track cooldowns, mechanics etc.

Structs

AttributeIter

An iterator over the variants of Attribute

BuffApplyEvent

Buff apply event.

BuffApplyKindIter

An iterator over the variants of BuffApplyKind

BuffCycleIter

An iterator over the variants of BuffCycle

BuffDamageEvent

Buff damage event.

BuffDamageResultIter

An iterator over the variants of BuffDamageResult

BuffFormula

Buff formula from an Event with StateChange::BuffFormula.

BuffInfo

Buff information from an Event with StateChange::BuffInfo.

BuffInitialEvent

Buff initial event.

BuffRemoveEvent

Buff remove event.

BuffRemoveIter

An iterator over the variants of BuffRemove

BuffRemoveKindIter

An iterator over the variants of BuffRemoveKind

BuffStackTypeIter

An iterator over the variants of BuffStackType

RawBuffFormula

Buff formula from an Event with StateChange::BuffFormula.

StackActiveEvent

Active buff stack change.

StackResetEvent

Buff stack reset.

Enums

Attribute

Attributes for buff formulas.

BuffApplyKind

Buff apply behavior.

BuffCategory

Buff info category after 13 December 2022.

BuffCategoryOld

Buff info category before 13 December 2022.

BuffCycle

Combat buff cycle.

BuffDamageResult

Buff damage tick results.

BuffRemove

Combat buff remove.

BuffRemoveKind

Kind of buff remove.

BuffStackType

Buff stacking behavior.