Module evtc::effect

Effect bindings & utilities.

Effects are visual effects rendered by the game client.

Structs

Effect

Effect information from an Event with StateChange::Effect.

EffectGUIDDeprecated

Effect information from an Event with StateChange::IdToGUID.

EffectOld

Effect information from an Event with StateChange::EffectOld.

EffectOrientation

Orientation of an effect.

Enums

EffectDuration

Duration of an effect in time or as a tracking id.

EffectLocation

Location of an effect.